Move FindStatus to find.rs
